[C++] Add optional support for unified function call syntax.
This makes equivalent f(x,y) and x.f(y), where f is marked as a candidate extension method. Candidacy is specified by qualifying the first parameter of a free-standing function with the 'this' keyword. Otherwise, there are no special considerations for these functions, and can still be called as a free-standing function. Rebased against llvm-11.0.1